This is a program on Hotel Management System developed using C++ Language.
C++ language is used to develop this Hotel Management application. Administrator will have all options and features in the application like book a room, see the customer details, to see the allotted room details and to edit the details of the customer by modifying it or can even delete the record.
In this system we will make extensive use of files system in C++.
We will have a login password system initially. In this system we will be having separate functions for:
•	Main menu to choose as per requirement.
•	Getting the information to allot a room.
•	Displaying customer details.
•	Display alloted rooms.
•	Modify a record.
•	Delete a record.
•	Calculate bill as per required
In the software developed separate functions will be there for each of the above points so that there is ample scope for adding more features in the near future.