This is small java project on Media Player developed by using NetBeans8.2 and SceneBuilder8.4.1
The main aim of designing and developing this Media Player primarily based Engineering project is to provide a media player that is capable of playing an mp4 file. Java language is used to develop this Media Player application. We have used JAVAFX , a feature of Netbeans8.2 . The FX in JAVAFX stands for Form eXtensible. It is a special feature in NetBeans for implementing the audio and visual effect. SceneBuilder8.4.1 is used for providing the interface in JAVAFX. Drag and Drop feature are available in SceneBuilder8.4.1 which generates the code in the NetBeans itself for the activities performed in SceneBuilder8.4.1.
Users will have the options and features in the application as to play the file, pause it, fast forward it, forward it, stop it, rewind it, fast rewind it, to exit. Volume Slider is also provided so as to increase or decrease the volume as per required. Time Slider is also there with the functionality of continuously moving until the video reaches to the end. Also, seeking the slider at any position and then playing the video from that particular location is also provided in the Time Slider functionality.